amend broken timestamps display
