+++ app/lib/database/views/runs_views.js
@@ -20,10 +20,10 @@
 const queryForRunsFields = `
         --p.name, 
         r.run_number, 
-        r.time_o2_start, 
-        r.time_o2_end, 
-        r.time_trg_start, 
-        r.time_trg_end,
+        extract( epoch from r.time_o2_start ) * 1000 as time_o2_start,
+        extract( epoch from r.time_o2_end ) * 1000 as time_o2_end,
+        extract( epoch from r.time_trg_start ) * 1000 as time_trg_start,
+        extract( epoch from r.time_trg_end ) * 1000 as time_trg_end,
         get_center_of_mass_energy(r.energy_per_beam, p.beam_type_id) as center_of_mass_energy, 
         r.fill_number,
         r.l3_current,
